services/api-gateway/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
httpx==0.27.0
PyJWT==2.9.0
pika==1.3.2
prometheus-fastapi-instrumentator==7.0.0
